add network report
